Unify proceeding visitors, remove casting from fork processing
