+++ workflows4s-core/src/main/scala/workflows4s/wio/WorkflowContext.scala
@@ -9,7 +9,7 @@ trait WorkflowContext { ctx: WorkflowContext =>
 
   type WIO[-In, +Err, +Out <: State] = workflows4s.wio.WIO[In, Err, Out, Ctx]
   object WIO extends AllBuilders[Ctx] {
-    type Branch[-In, +Err, +Out <: State]  = workflows4s.wio.WIO.Branch[In, Err, Out, Ctx]
+    type Branch[-In, +Err, +Out <: State]  = workflows4s.wio.WIO.Branch[In, Err, Out, Ctx, ?]
     type Interruption[+Err, +Out <: State] = workflows4s.wio.WIO.Interruption[Ctx, Err, Out]
     type Draft                             = WIO[Any, Nothing, Nothing]
     type Initial                           = workflows4s.wio.WIO.Initial[Ctx]
